UI changes to TechnicianAvailability and ScheduleBlock
UI:
- pass ticket object from accounts->scheduleservice to
scheduling->schedule
- add technician name on schedule block
- When I go into the availability screen I can see the availability for
all technicians,
with a name in each block.
- When I click on an existing availability record, I can only open it if
that record is for my user ID.
If not give an error message “That availability block does not belong to
you”.
- When I click to add an availability record, check to be sure I am a
technician.
If I am not, then give an error message that “You are a not a technician
and you cannot add availability”