Implemented - Time slots are 1 hour
Time slots are always 1 hour
1) If the user enters a certain number of time slots, then adjust the
ending time to be that number of hours from the starting time.
2) If the user resizes the block by dragging the handles and finishes
the resize to a 1/2 hour increment, then extend  to the next full hour
increment (note that the time block can start on at 30 minute increment,
like 8:30 am, but it always must end after an integer number of full
hours)
3)  If the user creates a block by drawing or resizging, then update the
"nubmer of slots" data field to reflect the number of 1 hour slots
included in the block.