Added re-schedule tab on scheduling snap-in
This would be a tab on the scheduler snap in.  Only tickets that are not
completed where the scheduled time has passed by 4 hours would display.
These are the tickets that need to be rescheduled.
When you click on the ticket, the ticket information comes up, with the
"save and schedule" button.  But there is a new button on the pop up for
"User Account"  and if you click this and NEW window opens and takes you
to that customers account screen.  When the user closes the account
screen, he still sees the ticket information.